requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
httpx==0.27.2
sse-starlette==2.1.3

# Data
sqlmodel==0.0.22
sqlalchemy==2.0.36
aiosqlite==0.20.0
greenlet==3.1.1
alembic==1.14.0

# Auth
pyjwt==2.10.0
email-validator==2.2.0
itsdangerous==2.2.0

# Multipart uploads + file parsing
python-multipart==0.0.17
pypdf==5.1.0

# Agent runtime
claude-agent-sdk==0.1.63
anthropic==0.39.0
openai==1.54.4

# Podcast pipeline (retained from V0)
elevenlabs==1.13.0
feedgen==1.0.0
pydub==0.25.1

# Scraping fallback (MCP tool)
beautifulsoup4==4.12.3
readability-lxml==0.8.1
lxml==5.3.0

# Utilities
rich==13.9.4
structlog==24.4.0
tenacity==9.0.0
